support type variables referencing type variables of enclosing classes
So far we support signatures like `Foo<S, T extends S>`, where one type parameter is bound by another type parameter. However it is possible, that type parameters of inner classes refer to type parameters of outer classes,
e.g. `class Foo<T> { class Inner<U extends T> {} }`.
This means that we not only need all constructed type parameters of the class under construction to finish the type arguments, but also all type parameters of all enclosing classes. Unfortunately enclosing class creation has been too late in the construction process for this, thus I have moved it to an earlier place, namely when type hierarchies are assembled. Also to ensure that the construction of the type parameters occurs in the correct order (after all we need the type parameters of the enclosing class constructed before we can process the type parameters of the inner class) I have added sorting by type name to the processing of imported classes. I realize that this is somewhat implicit to ensure the correct order this way (i.e. rely on the naming, that inner class names start with the outer classes name and thus sorting the class names lexicographically will result in the outer classes being processed first), but I still opted for this solution, because I think the complexity to integrate it into the current code base is low and the solution should not impact performance in a noticeable way. In the central place I added a comment and made the internal API clear by replacing the return type with `Sorted...` (at least to point out that the order is of relevance). Should this be broken in the future, tests will likely notice the problem, even though the symptom might unfortunately just be flaky tests.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>