Review: Use `assertThat(JavaClass).matches(Class<?>)`
It automatically does a lot of detail assertions that will ensure that the type has really been imported and is no stub (because `stub.getName()` still behaves like `Class.getName()`, but other properties will definitely be a mismatch in this case).
Also removed `isPresent()` checks. If there is a `.get()` in the next line anyway, the exception thrown by that call is IMHO good enough. Will make the assertion quicker to read to remove redundant assertions.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>