Review: Remove test for `lang` API
This is definitely the wrong place to test the rules API, since this is part of the `lang` layer (and the dependencies are `lang` -> `core`, not vice versa). Also I think the test is obsolete, because there are tests that meta-annotations principally work, so if we have tested that the transitive import works correctly on the core layer (i.e. the domain model is correctly created), there is no test gap left that needs to be filled.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>