optimize performance of `SourceCodeLocation`
Once again surprising, but for sufficiently many classes it has a noticeable performance impact to use `String.format(..)` instead of `StringBuilder` (which constant string concatenation will be compiled to in the byte code for all non-ancient JDK versions I know). Since we create the formatted source code location for every class in the import this scales linear with the number of classes.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>