Review: Added FIXME that `annotationStrategy` is obsolete as parameter
The strategy pattern is used to parameterize a specific part of behavior so the caller can decide on a "partial behavior" when they call the method. For example for `resolveInheritance` the caller can decide if they want to traverse the interface hierarchy or the class hierarchy by choosing the relevant strategy. For annotations this makes no sense, as it always is the same behavior and we should not prematurely parameterize this, if we do not have a specific use case in mind. Parameterizing part of the behavior via strategy will make the code more complex to understand, without an explicit need we should thus avoid it.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>