Review: parameterize test
Instead of calling `assertThatParametersOfMetaAnnotationAreImported` from two test methods that both import some element annotated with `SomeAnnotation` and then traversing the same hierarchy, we should actually parameterize the import. Because in the end we want to check two different inputs of a "thing" annotated with `@SomeAnnotation` (one time a class, one time a member) and assert for both cases that we correctly resolve the meta-annotation parameters.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>